Prototype test for job comment update.
This call is not working correctly. It succeeds but job data gets deleted in the process.
It seems the server may be taking the request body and using it to update the job
status. But this request has no body so it isn't supposed to do that.

Some server side debugging is required here.

@watanas